+++ CHANGELOG.md
@@ -6,6 +6,9 @@
   `completionCandidates` has been added to `Argument` and `Option` interfaces, and corresponding
   parameters have been added to the various implementation constructors, as well as the `convert`
   functions. You can use this to control the values autocomplete that will be suggested.
+- [`option().split()`](https://ajalt.github.io/clikt/api/clikt/com.github.ajalt.clikt.parameters.options/split.html),
+  and the corresponding
+  [`OptionWithValues.valueSplit`](https://ajalt.github.io/clikt/api/clikt/com.github.ajalt.clikt.parameters.options/-option-with-values/value-split.html).
 
 ### Changed
 - If multiple `--` tokens are present on the command line, all subsequent occurrences after the first are now